code cleanup: apply eslint and reorganize some functions
