+++ package.json
@@ -47,6 +47,7 @@
     "uuid": "13.0.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.38.0",
     "@storybook/addon-docs": "9.1.7",
     "@storybook/addon-essentials": "8.6.14",
     "@storybook/addon-interactions": "8.6.14",
@@ -68,10 +69,12 @@
     "eslint-plugin-react-hooks": "5.2.0",
     "eslint-plugin-react-refresh": "0.4.20",
     "eslint-plugin-storybook": "9.1.7",
+    "globals": "^16.4.0",
     "jsdom": "27.0.0",
     "prop-types": "15.8.1",
     "socket.io": "4.8.1",
     "storybook": "9.1.7",
+    "typescript-eslint": "^8.46.2",
     "vite": "7.1.6",
     "vitest": "3.2.4"
   },
